backend/requirements.txt

fastapi
uvicorn[standard]
sse-starlette
python-multipart
opencv-python-headless
numpy
python-chess
